+++ models/access_db.py
@@ -12,6 +12,12 @@ class AccessModel(BasicModel):
 
         self.logger.info('{} start'.format(self.model))
 
+    def get(self, page=1):
+        data_list = self.collection.find().sort('timestamp', -1)
+        get_page = Page(page)
+        paging, data_list = get_page.paginate(data_list, collection=self.collection)
+        return paging, data_list
+
     def get_by_ticket(self, ticket, page=1):
         result = self.db['tickets'].find_one({'ticket': ticket})
         if result:
@@ -20,4 +26,11 @@ class AccessModel(BasicModel):
             data_list = []
         get_page = Page(page)
         paging, data_list = get_page.paginate(data_list, collection=self.collection)
+        return paging, data_list
+    
+    def get_by_site(self, site, page=1):
+        query = {'host': site}
+        data_list = self.collection.find(query).sort('timestamp', -1)
+        get_page = Page(page)
+        paging, data_list = get_page.paginate(data_list, query=query, collection=self.collection)
         return paging, data_list